Move `SystemNativeUtils` to more generic location
